Build the static component structure for timesheet entries
