+++ components/timesheet-page/timesheet/timesheet-entry/TimesheetEntry.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+import './timesheet-entry.scss';
+
+const TimesheetEntry = () => (
+  <div className="timesheet-entry-wrapper timesheet-entry-wrapper__good">
+    <p className="timesheet-entry__employer">
+      Port of Rotterdam
+    </p>
+    <p className="timesheet-entry__date">
+      09:00 - 17:00
+    </p>
+  </div>
+);
+
+export default TimesheetEntry;
