+++ components/timesheet-page/timesheet/timesheet-entry/timesheet-entry.scss
@@ -1,0 +1,25 @@
+@import "../../../../shared/styles/main";
+
+.timesheet-entry-wrapper {
+  border: solid 1px $color-accent-2;
+  display: flex;
+  justify-content: space-between;
+  padding: 30px 16px 33px;
+  // padding-bottom: 33;
+
+  &__good {
+    border-left: solid 4px $color-primary;
+  }
+}
+
+.timesheet-entry {
+  &__employer {
+    color: $color-accent-5;
+    font-size: 16px;
+  }
+
+  &__date {
+    color: $color-accent-5;
+    font-size: 16px;
+  }
+}
