+++ components/timesheet-page/timesheet/timesheet.scss
@@ -1,0 +1,12 @@
+@import "../../../shared/styles/main";
+
+.timesheet-wrapper {
+  background-color: $color-monochrome-1;
+  // background-color: $color-accent-2;
+  margin: 20px auto 0;
+  width: 90%;
+
+  &__date {
+    color: $color-accent-6;
+  }
+}
