+++ shared/styles/_variables.scss
@@ -19,6 +19,8 @@ $color-accent: #1166a5;
 $color-accent-2: #f5f8fa;
 $color-accent-3: #e6eaee;
 $color-accent-4: #68768c;
+$color-accent-5: #354052;
+$color-accent-6: #7f8fa4;
 
 $color-error: #f60000;
 
@@ -38,6 +40,7 @@ $color-accent-alt-5: #f2f4f7;
 $color-accent-alt-6: #35ac45;
 $color-accent-alt-7: #39b54a;
 $color-accent-alt-8: #34aa44;
+
 
 
 /* color specific */
